fix(knowledge): count coercion + working re-sync button
Two bugs in the live Knowledge tab caught by Cam's eyes-on review:

1. Status pill showed green "Indexed" even when 0 pages were indexed.
   Cause: COUNT(...) returns Postgres bigint which the pg driver serializes
   as a string ("0"), not a number. The `sql<number>` annotation was a lie.
   Strict equality (pagesIndexed === 0) silently failed so the WebsiteContent
   component fell through to the default "Indexed" green branch.
   Fix: cast COUNT(...) and DISTINCT counts to ::int in SQL across page.tsx
   (page + file queries) and indexer.ts (getIndexingStatus). Same fix
   applied to the file-list chunks count for consistency.

2. While I was in indexer.ts: getIndexingStatus also had a second bug \u2014 the
   "most recent sync" query ordered lastSyncedAt ASC and took LIMIT 1, so it
   returned the OLDEST timestamp instead of the newest. Folded into the same
   grouped query with MAX(lastSyncedAt).

3. Re-sync button is now wired up (was a dead disabled control). New endpoint
   POST /api/knowledge/site/resync wipes the tenant's page rows and triggers
   a fresh indexTenantSite() via after() so it survives Vercel's serverless
   response. Added scripts/qa-resync.mjs as a regression smoke test \u2014
   verified end-to-end against the live Supabase backend: 3 pages indexed \u2192
   wipe \u2192 3 pages re-indexed. Delta detection / age-warning UX still owned by
   CON-86; this is the minimum so the button isn't lying to users.

Gates: tsc clean, eslint clean, npm run build clean.

Refs: CON-85 follow-up (count + sync timestamp bugs); pre-empts the
re-sync MVP slice of CON-86.

+++ src/app/dashboard/knowledge/page.tsx
@@ -11,7 +11,8 @@ export default async function KnowledgePage() {
   const tenant = await getCurrentTenant();
   if (!tenant) redirect("/onboarding");
 
-  // Fetch files with chunk counts
+  // Fetch files with chunk counts. Same bigint-as-string trap as above — cast
+  // to int so the file-list UI sees a real number for chunksIndexed.
   const files = await db
     .select({
       id: knowledgeFiles.id,
@@ -22,7 +23,7 @@ export default async function KnowledgePage() {
       uploadedAt: knowledgeFiles.uploadedAt,
       indexedAt: knowledgeFiles.indexedAt,
       errorMessage: knowledgeFiles.errorMessage,
-      chunksIndexed: sql<number>`COUNT(${knowledgeItems.id})`.as(
+      chunksIndexed: sql<number>`COUNT(${knowledgeItems.id})::int`.as(
         "chunks_indexed"
       ),
     })
@@ -32,10 +33,13 @@ export default async function KnowledgePage() {
     .groupBy(knowledgeFiles.id)
     .orderBy(sql`${knowledgeFiles.uploadedAt} DESC`);
 
-  // Fetch website indexing stats
+  // Fetch website indexing stats. NOTE: COUNT(...) returns Postgres bigint which
+  // the `pg` driver serializes as a string ("0", "42"). Cast to integer in SQL
+  // so the typed result is actually a number — otherwise downstream strict
+  // equality (pagesIndexed === 0) silently fails and the status pill misreads.
   const [websiteStats] = await db
     .select({
-      pagesIndexed: sql<number>`COUNT(DISTINCT ${knowledgeItems.sourceUrl})`.as(
+      pagesIndexed: sql<number>`COUNT(DISTINCT ${knowledgeItems.sourceUrl})::int`.as(
         "pages_indexed"
       ),
       lastSynced: sql<Date | null>`MAX(${knowledgeItems.lastSyncedAt})`.as(
